+++ packages/macros/tests/glimmer/fail-build.test.ts
@@ -4,6 +4,7 @@ import { MacrosConfig } from '../..';
 describe(`macroFailBuild`, function() {
   templateTests(function(transform: (code: string) => string, config: MacrosConfig) {
     config.setOwnConfig(__filename, { failureMessage: 'I said so' });
+    config.finalize();
 
     test('it can fail the build, content position', () => {
       expect(() => {